chore: move things into the hooks/ etc structure
