--- dropbox-hacking-photo-manager-client/src/context/countsByDateContext.tsx
@@ -1,5 +1,0 @@
-import { useAdditionalFeeds } from "./rx/additionalFeeds";
-import { useLatestValue } from "./rx/useLatestValue";
-
-export const useCountsByDate = () =>
-  useLatestValue(useAdditionalFeeds()?.countsByDate);
