+++ dropbox-hacking-photo-manager-client/src/hooks/legacyRxFeeds/countsByDateContext.tsx
@@ -1,0 +1,5 @@
+import { useAdditionalFeeds } from "./additionalFeeds";
+import { useLatestValue } from "../useLatestValue";
+
+export const useCountsByDate = () =>
+  useLatestValue(useAdditionalFeeds()?.countsByDate);
