+++ dropbox-hacking-photo-manager-client/src/pages/legacy/photo/index.tsx
@@ -3,8 +3,8 @@ import * as React from "react";
 import { useEffect, useMemo, useState } from "react";
 
 import logRender from "@lib/logRender";
-import { useLatestValue } from "@context/rx/useLatestValue";
-import { useAdditionalFeeds } from "@context/rx/additionalFeeds";
+import { useLatestValue } from "@hooks/useLatestValue";
+import { useAdditionalFeeds } from "@hooks/legacyRxFeeds/additionalFeeds";
 import SamePageLink from "@components/samePageLink";
 import EditableTextField from "@components/editableTextField";
