+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/day/files.tsx
@@ -3,10 +3,10 @@ import React, { useEffect, useMemo } from "react";
 import logRender from "@lib/logRender";
 import Navigate from "@components/navigate";
 import type { DayFilesResult } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
-import { useLatestValueFromServerFeed } from "../useLatestValueFromServerFeed";
+import { useLatestValueFromServerFeed } from "../../../hooks/useLatestValueFromServerFeed";
 import EditableTextField from "@components/editableTextField";
 import FilesTable from "./filesTable";
-import { useIdentity } from "./useIdentity";
+import { useIdentity } from "../../../hooks/useIdentity";
 
 const NGDayFiles = ({ date }: { date: string }) => {
   console.log("NGDayFiles", useIdentity());
