+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/exifExplorer/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 
 import logRender from "@lib/logRender";
 import Navigate from "@components/navigate";
-import { useLatestValueFromServerFeed } from "../useLatestValueFromServerFeed";
+import { useLatestValueFromServerFeed } from "../../../hooks/useLatestValueFromServerFeed";
 import type { ExifExplorerType } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
 
 type Counts = ExifExplorerType["tagCounts"][number][1];
