+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/fileRev/index.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from "react";
 import logRender from "@lib/logRender";
 import Navigate from "@components/navigate";
 import type { FileRevResult } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
-import { useLatestValueFromServerFeed } from "../useLatestValueFromServerFeed";
+import { useLatestValueFromServerFeed } from "../../../hooks/useLatestValueFromServerFeed";
 import SamePageLink from "@components/samePageLink";
 
 const NGFileRev = ({ rev }: { rev: string }) => {
