+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/fsck/index.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from "react";
 import logRender from "@lib/logRender";
 import Navigate from "@components/navigate";
 import type { FsckType } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
-import { useLatestValueFromServerFeed } from "../useLatestValueFromServerFeed";
+import { useLatestValueFromServerFeed } from "../../../hooks/useLatestValueFromServerFeed";
 
 const Fsck = () => {
   const latestValue = useLatestValueFromServerFeed<FsckType>({
