+++ dropbox-hacking-photo-manager-client/webpack.config.ts
@@ -44,7 +44,6 @@ const config: webpack.Configuration = {
   resolve: {
     alias: {
       "@components/*": path.resolve(__dirname, "src/components/*"),
-      "@context/*": path.resolve(__dirname, "src/context/*"),
       "@hooks/*": path.resolve(__dirname, "src/hooks/*"),
       "@lib/*": path.resolve(__dirname, "src/lib/*"),
       "@pages/*": path.resolve(__dirname, "src/pages/*"),
